HydraBoot/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pip==19.2.3
PyYAML==5.3.1
setuptools==41.2.0
Werkzeug==1.0.1
